# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CertChain - Blockchain Certificate Authentication and Issuer Validation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IssuerType = (typeof IssuerType)[keyof typeof IssuerType];

export const IssuerType = {
  university: "university",
  college: "college",
  school: "school",
  organization: "organization",
  government: "government",
} as const;

export type IssuerStatus = (typeof IssuerStatus)[keyof typeof IssuerStatus];

export const IssuerStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Issuer {
  id: number;
  name: string;
  type: IssuerType;
  country: string;
  address: string;
  publicKeyHash: string;
  status: IssuerStatus;
  votesFor: number;
  votesAgainst: number;
  totalCertificates: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateIssuerBodyType =
  (typeof CreateIssuerBodyType)[keyof typeof CreateIssuerBodyType];

export const CreateIssuerBodyType = {
  university: "university",
  college: "college",
  school: "school",
  organization: "organization",
  government: "government",
} as const;

export interface CreateIssuerBody {
  name: string;
  type: CreateIssuerBodyType;
  country: string;
  address: string;
}

export type UpdateIssuerStatusBodyStatus =
  (typeof UpdateIssuerStatusBodyStatus)[keyof typeof UpdateIssuerStatusBodyStatus];

export const UpdateIssuerStatusBodyStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateIssuerStatusBody {
  status: UpdateIssuerStatusBodyStatus;
}

export type VoteOnIssuerBodyVote =
  (typeof VoteOnIssuerBodyVote)[keyof typeof VoteOnIssuerBodyVote];

export const VoteOnIssuerBodyVote = {
  approve: "approve",
  reject: "reject",
} as const;

export interface VoteOnIssuerBody {
  vote: VoteOnIssuerBodyVote;
  validatorName: string;
}

export type CertificateStatus =
  (typeof CertificateStatus)[keyof typeof CertificateStatus];

export const CertificateStatus = {
  valid: "valid",
  revoked: "revoked",
  expired: "expired",
} as const;

export interface Certificate {
  id: number;
  certificateHash: string;
  holderName: string;
  holderEmail: string;
  issuerId: number;
  issuerName: string;
  degree: string;
  field: string;
  grade: string;
  issuedDate: string;
  expiryDate?: string | null;
  status: CertificateStatus;
  transactionHash: string;
  blockNumber: number;
  createdAt: string;
}

export interface IssueCertificateBody {
  holderName: string;
  holderEmail: string;
  issuerId: number;
  degree: string;
  field: string;
  grade: string;
  issuedDate: string;
  expiryDate?: string | null;
}

export type UpdateCertificateStatusBodyStatus =
  (typeof UpdateCertificateStatusBodyStatus)[keyof typeof UpdateCertificateStatusBodyStatus];

export const UpdateCertificateStatusBodyStatus = {
  revoked: "revoked",
} as const;

export interface UpdateCertificateStatusBody {
  status: UpdateCertificateStatusBodyStatus;
}

export interface VerifyCertificateBody {
  certificateHash: string;
  verifierName: string;
  verifierOrganization: string;
}

export interface VerificationResult {
  found: boolean;
  valid: boolean;
  certificate?: Certificate | null;
  searchTimeMs: number;
  bloomFilterHit: boolean;
  message: string;
}

export type VerificationResultProperty =
  (typeof VerificationResultProperty)[keyof typeof VerificationResultProperty];

export const VerificationResultProperty = {
  valid: "valid",
  revoked: "revoked",
  not_found: "not_found",
} as const;

export interface Verification {
  id: number;
  certificateHash: string;
  verifierName: string;
  verifierOrganization: string;
  result: VerificationResultProperty;
  searchTimeMs: number;
  bloomFilterHit: boolean;
  createdAt: string;
}

export interface DashboardStats {
  totalCertificates: number;
  validCertificates: number;
  revokedCertificates: number;
  totalIssuers: number;
  approvedIssuers: number;
  pendingIssuers: number;
  totalVerifications: number;
  successfulVerifications: number;
  avgSearchTimeMs: number;
  bloomFilterEfficiency: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  certificate_issued: "certificate_issued",
  certificate_revoked: "certificate_revoked",
  issuer_approved: "issuer_approved",
  issuer_rejected: "issuer_rejected",
  verification_success: "verification_success",
  verification_failed: "verification_failed",
  issuer_registered: "issuer_registered",
  vote_cast: "vote_cast",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  entityName: string;
  createdAt: string;
}

export interface TrendPoint {
  date: string;
  verifications: number;
  successful: number;
}

export type ListIssuersParams = {
  status?: ListIssuersStatus;
  search?: string;
};

export type ListIssuersStatus =
  (typeof ListIssuersStatus)[keyof typeof ListIssuersStatus];

export const ListIssuersStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type ListCertificatesParams = {
  issuerId?: number;
  status?: ListCertificatesStatus;
  search?: string;
};

export type ListCertificatesStatus =
  (typeof ListCertificatesStatus)[keyof typeof ListCertificatesStatus];

export const ListCertificatesStatus = {
  valid: "valid",
  revoked: "revoked",
  expired: "expired",
} as const;

export type ListVerificationsParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};